chore(lexer): define the lexer output
Since the parser will mostly iterate on TokenKinds, we use a "Struct of arrays" layout.

The TokenInfo struct has metadata around TokenKind that will let us get the literal
from a []byte. This will be feasible by indexing with a combination of LineIdx and
Column.

The LineInfo struct records the offset in the []byte and the length of a given line.