feat(lexer): lex symbols
We use a stateFn as a combination of state and action. It executes
an action and returns the next state.

We also compute the line infos ahead of time before lexing the tokens.
This should be more performant than handling another state in the hot loop (Lex).

See-Also: https://go.dev/talks/2011/lex.slide