test(parser): tests "à la" tree-sitter
For testing, having the parse tree as text is much more convenient.
As such, we introduce a way of testing similar to tree-sitter. This
means we can have a file containing:

```
================================================================================
TEST NAME
================================================================================

INPUT TO PARSER

--------------------------------------------------------------------------------

EXPECTED OUTPUT OF PARSER
```

From that file, we can generate a ParseTestCase, and test against
the actual output of the parser.

See-Also: https://github.com/tree-sitter