feat(linker): add basic multiset discrimination
First, we add kinds to ParseTree nodes. Since there was already a hole
in the data alignment of the Nodes, we get to store the kind for "free".

Second, we go through all the declaration in the linker and record the
names of each message/enum/service. Note that it is important to handle
nested messages and enums.

Finally, we add the types names to a multiset. This is useful because
after sorting (Three-way string quicksort) it, all the declarations and
usages of a type will be next to each other. This will let us find
duplicate declarations or usage of non declared types later on.

See-Also: https://www.informit.com/articles/article.aspx?p=2180073&seqNum=4