feat(typecheck): change algorithm to check decls and refs
Instead of using a basic multiset discrimination, we now separate the refs and the decls into two arrays.
The decls array contains the declaration name preceded by the pkg + scope (e.g. my.pkg.A.B). And the refs
array contains a representation of the refs scope and the ref name. For example, if we look for A.B in
message C under package a.b, the string representing the ref will be .a.b.C]A.B . This means that we can
look for A.B in a.b.C, a.b, a, or ".". The refs strings can also contain a mininmum scope part. Fo example,
.a.b[C]A.B means we can only check for A.B in a.b.C and a.b .